requirements.txt
asgiref==3.3.1
Django==3.1.7
django-phonenumber-field==5.0.0
djangorestframework==3.12.2
install==1.3.4
phonenumbers==8.12.18
Pillow==8.1.0
pytz==2021.1
sqlparse==0.4.1
